Fixed the output to float instead of Integer

+++ 0x02-python_async_comprehension/0-async_generator.py
@@ -1,11 +1,11 @@
-#!/usr/bin/env python3.7
+#!/usr/bin/env python3
 """ Async Generator """
 import asyncio
 from random import uniform
 from typing import Generator
 
 
-async def async_generator() -> Generator[int, None, None]:
+async def async_generator() -> Generator[float, None, None]:
     """ Async Generator """
     for i in range(10):
         yield uniform(0, 10)
